backend/pyproject.toml
[project]
name = "phantom-mask-backend"
version = "0.1.0"
description = "A Project for Phantom Mask"
authors = ["0xJasonChien <jason891588@gmail.com>"]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-csp>=4.0",
    "django-filter>=25.1",
    "django-simple-captcha>=0.6.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "pre-commit>=4.2.0",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.9.1",
]

[tool.ruff]
exclude = [
    "migrations",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["E501", "D", "PLC0415", "ARG002"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"


[tool.black]
skip-string-normalization = true
